third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryServices/gen-garbage-collection-ran-during-error-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  let promise;
  return getEmptyHealthThermometerDevice()
      .then(({device}) => {
        promise = assert_promise_rejects_with_message(
          device.gatt.getPrimaryServices('health_thermometer'),
          new DOMException(
            'GATT Server is disconnected. ' +
            'Cannot retrieve services. ' +
            '(Re)connect first with `device.gatt.connect`.',
            'NetworkError'));
        // Disconnect called to clear attributeInstanceMap and allow the
        // object to get garbage collected.
        device.gatt.disconnect();
      })
      .then(runGarbageCollection)
      .then(() => promise);
}, 'Garbage Collection ran during a getPrimaryServices call that failed. ' +
   'Should not crash.');

</script>
